New features
- Changed BASE_DIR determintation logic
- Added possibility to deploy dumps from current folder

+++ dockerup.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-BASE_DIR="$(dirname "$0")";
+BASE_DIR="$(dirname `readlink "$0"`)";
 
 source "$BASE_DIR/config.sh";
 if [ -f "$BASE_DIR/config_custom.sh" ]; then
@@ -66,11 +66,76 @@ sshfs "$TICKET_NUMBER":/var/www/html/ "$CONTAINER_PATH/src/" -ocache=no;
 DOMAIN="$TICKET_NUMBER.$CONTAINERS_DOMAIN_SUFFIX"
 sudo sh -c "echo '$IP_ADDRESS     $DOMAIN' >> /etc/hosts"
 
+M2_DUMPS_DEPLOYED=0
+if [ "$MAGENTO_VERSION" = "m2" ] && [ "$DOCKER_IMAGE_NAME" = "base" ]
+then
+
+    echo "Creating .m2install.conf file..."
+    cp "$BASE_DIR/template/.m2install.conf" "$CONTAINER_PATH/.m2install.conf"
+    sed -i '' s/%domain%/"$DOMAIN"/g $CONTAINER_PATH/.m2install.conf;
+    scp $CONTAINER_PATH/.m2install.conf $TICKET_NUMBER:/var/www/html
+    echo "Created!"
+
+    echo "Finding code dump..."
+    codeDumpFilename=$(find . -maxdepth 1 -name '*.tbz2' -o -name '*.tar.bz2' | head -n1)
+    if [ "${codeDumpFilename}" == "" ]
+    then
+        codeDumpFilename=$(find . -maxdepth 1 -name '*.tar.gz' | grep -v 'logs.tar.gz' | head -n1)
+    fi
+    if [ ! "$codeDumpFilename" ]
+    then
+        codeDumpFilename=$(find . -maxdepth 1 -name '*.tgz' | head -n1)
+    fi
+    if [ ! "$codeDumpFilename" ]
+    then
+        codeDumpFilename=$(find . -maxdepth 1 -name '*.zip' | head -n1)
+    fi
+
+    if [ "$codeDumpFilename" != "" ]
+    then
+        echo "Code dump $codeDumpFilename was copied to container!"
+        scp "$codeDumpFilename" $TICKET_NUMBER:/var/www/html
+    else
+        echo "Code dump was not found!"
+    fi
+
+    echo "Finding db dump..."
+    dbdumpFilename=$(find . -maxdepth 1 -name '*.sql.gz' | head -n1)
+    if [ ! "$dbdumpFilename" ]
+    then
+        dbdumpFilename=$(find . -maxdepth 1 -name '*_db.gz' | head -n1)
+    fi
+    if [ ! "$dbdumpFilename" ]
+    then
+        dbdumpFilename=$(find . -maxdepth 1 -name '*.sql' | head -n1)
+    fi
+
+    if [ "$dbdumpFilename" != "" ]
+    then
+        echo "Database dump $codeDumpFilename was copied to container"
+        scp "$dbdumpFilename" $TICKET_NUMBER:/var/www/html
+    else
+        echo "Code dump was not found!"
+    fi
+
+    if [ "$codeDumpFilename" != "" ] && [ "$dbdumpFilename" != "" ]
+        echo "Code and database dumps were found. Staring m2install tool..."
+        ssh $TICKET_NUMBER "cd /var/www/html; m2install.sh --force"
+        M2_DUMPS_DEPLOYED=1
+    then
+        echo "Code and database dumps were not found. m2install will not run automatically!"
+    fi
+fi
+
 # Set base url
-ssh $TICKET_NUMBER "mysql -umagento -p123123q magento -e \"UPDATE core_config_data SET value=\\\"http://$DOMAIN/\\\" WHERE path=\\\"web%url\\\";\""
+if [ "$M2_DUMPS_DEPLOYED" != "1" ]
+then
+    ssh $TICKET_NUMBER "mysql -umagento -p123123q magento -e \"UPDATE core_config_data SET value=\\\"http://$DOMAIN/\\\" WHERE path=\\\"web%url\\\";\""
+fi
 
 # Cleaning and static deploy
-if [ "$MAGENTO_VERSION" = "m2" ]; then
+if [ "$MAGENTO_VERSION" = "m2" ] && [ "$M2_DUMPS_DEPLOYED" != "1" ]
+then
     ssh $TICKET_NUMBER "sudo rm -Rf /var/www/html/var/*; rm -Rf /var/www/html/pub/static/frontend; rm -Rf /var/www/html/pub/static/adminhtml; rm -Rf /var/www/html/pub/static/_requirejs; cd /var/www/html/; php bin/magento setup:static-content:deploy"
 fi
